requirements.txt
nltk
werkzeug
flask
py-postgresql
rake_nltk
pymystem3
